Edited Main to be 2 files. And edited the main in the Main to reference to the CommandlineParser file.
